% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_specplot.R
\name{batch_specplot}
\alias{batch_specplot}
\title{batch_specplot}
\usage{
batch_specplot(spec_files, file_path = 'C://Users//eric_fortune//Desktop', name = 'birds', width = 100, height = 100)
}
\arguments{
\item{input_array}{2-D array of frequency data already broken into chunks}

\item{Fs}{Sample frequency of audio data}

\item{file_path}{Path to save image files}

\item{name}{Base name of files}

\item{width}{Width in pixels of output files. Defaults to 299}

\item{height}{Height in pixels of output files. Defaults to 299}

\item{...}{Other arguments to pass to specplot function. See specplot from DFWE for arguments. NOTE: amp_range must be
specified here}
}
\value{
This function outputs a multidimensional array of rgb images
}
\description{
Plot batches of spectrograms from arrays of chunks. Uses arrays output from the training_cut function.
Will output images of spectrograms for use in neural network.
}
\examples{
batch_specplot(spec_files, file_path = 'C://Users//eric_fortune//Desktop', name = 'birds', width = 100, height = 100)

}
