% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domplot.R
\name{domplot}
\alias{domplot}
\title{domplot}
\usage{
domplot(wav_file, Fs, nfft, amp_range, dfcol, bat = TRUE, plotreg = FALSE, ...)
}
\arguments{
\item{wav_file}{A .wav file or list of frequency data}

\item{Fs}{Sample rate, optional if included in .wav file}

\item{nfft}{Default value 128 when bat = TRUE, when bat = FALSE the default is 512}

\item{amp_range}{Changes default mapping of colors to decible range}

\item{dfcol}{Sets color for dominant frequency points}

\item{bat}{Defaults to TRUE, a series of default values tuned to produce nice plots from bat data}

\item{plotreg}{Default FALSE. Plot Linear Regression line. May not be a good fit on all syllables.}

\item{...}{Pass parameters on to specplot function !!EXCEPT amp_range!!}
}
\description{
Plots the dominate frequency of a specrogram. Uses seeWave dfreq function. Will plot directly on top of spectrogram. This function
will plot and automatically output the matrix result of dfreq. This allows some other analysis to be done, for instance taking
minimum/maximum or generating a linear regression.
}
\examples{
domplot(bat_syllables$syl1, f = 250000, dfcol = "White")

}
