% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seesyl.R
\name{seesyl}
\alias{seesyl}
\title{seesyl}
\usage{
slop(zfinch_data, zf_syllables)
}
\arguments{
\item{wav_file}{A .wav file or list of frequency data}

\item{syllable_data}{A data frame produced by the sepsyll function}

\item{Fs}{Sample rate, optional if included in .wav file}

\item{nfft}{Default value 128 when bat = TRUE, when bat = FALSE the default is 512}

\item{amp_range}{Changes default mapping of colors to decible range}

\item{bat}{Defaults to TRUE, a series of default values tuned to produce nice plots from bat data}

\item{...}{Pass parameters on to specplot function !!EXCEPT amp_range!!}
}
\value{
A plot containing marked syllables and dominate frequencies
}
\description{
Graphs spectrogram of bat or bird sounds and marks beginning and end of syllables. slop can also plot dominate frequency on
top of the spectrogram.
}
\examples{
slop(wav_file = bat_sounds, syllable_data = bat_syl)

}
