% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slop.R
\name{slop}
\alias{slop}
\title{slop}
\usage{
slop(zfinch_data, zf_syllables)
}
\arguments{
\item{wav_file}{A .wav file or list of frequency data}

\item{syllable_data}{A data frame produced by the sepsyll function}

\item{Fs}{Sample rate, optional if included in .wav file}

\item{nfft}{Default value 128 when bat = TRUE, when bat = FALSE the default is 512}

\item{dom_freq}{defaults to TRUE, includes dominate frequency plot}

\item{bat}{Defaults to TRUE, a series of default values tuned to produce nice plots from bat data}
}
\value{
A plot containing marked syllables and dominate frequencies
}
\description{
Graphs spectrogram of bat or bird sounds and marks beginning and end of syllables. slop can also plot dominate frequency on
top of the spectrogram.
}
\examples{
slop(wav_file = bat_sounds, syllable_data = bat_syl)

}
