% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_cut.R
\name{training_cut}
\alias{training_cut}
\title{training_cut}
\usage{
training_cut(audio_data, fs, chunk_length)
}
\arguments{
\item{audio_data}{Audio data as either a wav file of a list of frequencies}

\item{Fs}{Sample rate, does not need to be included if it is encoded in a wav file}

\item{chunk_length}{Length of chunk in seconds, defaults to 1 second}
}
\value{
An dataframe of chunks containing raw audio data
}
\description{
This function cuts up preprocessed recordings into even chunks.
}
\examples{
traning_cut(bcw_full, chunk_length = 0.5)

}
