% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specplot.R
\name{specplot}
\alias{specplot}
\title{specplot
This function will plot spectrograms. It works with either a wav file read by tuneR or frequency data.
When using frequency data alone, you need to specify a sample frequency. This function relies on the
imagep() function from oce. specplot() uses custom color palettes with 30 colors. Most parameters have defaults
and do not need to be specified by the user.}
\usage{
specplot = function(freq_data, Fs, nfft, wl, ovlp, normal = TRUE, amp_range, color, amp_value = FALSE)
}
\arguments{
\item{freq_data}{Wav file or frequency data}

\item{Fs}{Sample rate, not necessary if provided by .wav file}

\item{nfft}{Defaults to 512}

\item{wl}{Defaults to 1/2 nfft}

\item{ovlp}{Overlap, give in percent. Defaults to 50%}

\item{normal}{Normalize. Defaults to TRUE}

\item{amp_range}{DB range}

\item{color}{Choose from 4 present color palettes; 1 = reverse heat (default), 2 = reverse greyscale, 3 = greyscale, 4 = heat}

\item{amp_value}{Display min and max dB. Defaults to FALSE}
}
\value{
A plot of a spectrgram using the default R graphics
}
\description{
specplot
This function will plot spectrograms. It works with either a wav file read by tuneR or frequency data.
When using frequency data alone, you need to specify a sample frequency. This function relies on the
imagep() function from oce. specplot() uses custom color palettes with 30 colors. Most parameters have defaults
and do not need to be specified by the user.
}
\examples{
specplot(zfinch_data, ovlp = 90, color = 2, amp_value = TRUE)

}
